Use mimemagic 0.3.9
This version of the gem is licenced under MIT, so it's fine for us to
continue using.

It depends on a locally installed MIME database rather than it coming
bundled with the gem.

This seems to be the approach that other development teams in MOJ are
using, so I'm following their lead. See Slack thread here:
https://mojdt.slack.com/archives/C7KCVJZSQ/p1616750434063200

### Local development

On local development machines, this means we'll need to install the
homebrew package `shared-mime-info`. I've updated the `Brewfile` to
include this as a project dependency – so developers will need to run
`brew bundle` in their terminal to install it.

### Docker builds

`shared-mime-info` is already included in the `ruby:2.6.5-stretch` image
which we build upon, so it's already accessible to the gem. No further
action is required for this to work.