Upgrade versions of Ruby, Rails and Debian
This commit upgrades the application to run on:
- Debian Bullseye (was Debian Stretch)
- Ruby 2.6.8 (was 2.6.5)
- Rails 6.1.4.1 (was 6.1.2.1)
- Bundler 2 (was Bundler 1)

This brings the service's software dependencies up-to-date. There is
still a major Ruby upgrade to be performed (to version 3), but I've left
that out for now.

The main driving force behind this upgrade is to make sure we're not
unnecessarily running out of date versions for important but often
overlooked dependencies (e.g. the Debian docker base image used)

I've also removed the `package.json` file since this project doesn't
have any JavaScript dependencies. This is a backend API service which
has no need for frontend assets, nor does it have any Node.js
dependencies.