MO-1518 Fix the 'multiple' endpoint
Ticket: https://dsdmoj.atlassian.net/browse/MO-1518

This is a second attempt to fix an issue with the `multiple` endpoint. Previously a similar approach was taken but PR reverted as the API stopped working, which I believe must have been some unrelated event, not produced by the code changes.

More info:
https://mojdt.slack.com/archives/C03N99X465Q/p1711029843538929

In summary, the `multiple` endpoint was sometimes returning incorrect CNL for some offenders, and the culprit was the DB query scoping by `active: true`, meaning it would be returning ONLY records that are active, and from those, picking the most recent one for each of the offenders passed.

However the most recent record might have been deactivated (as in having no CNL), but the scoping was not returning it.

This PR takes the previous attempt and does the same thing to try to fix the problem, just a slightly different code.