MO-1128 Reintroduce running tests on CI
Ticket: https://dsdmoj.atlassian.net/browse/MO-1128

As part of PR #51 tests on CI were disabled due to not having necessary secrets available to request a valid jwt token from hmpps auth service.

This is bad practice so we want to reintroduce running tests on CI. In order to do this I'm mocking the auth calls if we are running tests on CI, whereas in local envs it will continue as before, requesting real jwt tokens.

As part of this work I've done a cleanup of any mention to the token "scope" because it was not used in the code, we are only using the roles.