chore: Dependabots and rubocop linting
Minor bump to some gems to close dependabots. Fixed some new rubocop offences.

Update `dependabot.yml` to include docker, and to only raise PRs for security versions, and not for version updates. This will keep the repo updated and notify us whenever there are security issues but not flood us with minor, unnecessary versions. Minor bumps can be bundled together and done periodically as part of the general maintenance of the app.