Quicker pipeline on merge to main
Removed the need to run tests and linters on `main` branch, given our feature branches / pull requests enforce the branch to be up to date with main before being able to merge. So if the tests/linters are ok in the PR, we are certain there will be no failures upon merge to main. Removing these from running on main branch will cut some time to the deploys particularly in MPC where there are many tests and a bit slow.