MO-1529 Update rails to 7.0.x
Ticket: https://dsdmoj.atlassian.net/browse/MO-1529

I'm separating the update in 2 parts to reduce potential risk of introducing regressions.

First part will update rails to the latest patch in the 7.0 branch, and the second part will update to the latest patch in the 7.1 branch.

As part of this PR I've also removed `timecop` gem in favor of the rails included time helpers.